+++ ru.taximaxim.codekeeper.ui/src/ru/taximaxim/codekeeper/ui/UIConsts.java
@@ -126,6 +126,7 @@ public interface UIConsts {
         String SCRIPT_FROM_SELECTED_OBJS = "prefScriptFromSelectedObjs"; //$NON-NLS-1$
         String DATA_MOVEMENT_MODE = "prefDataMovementMode"; //$NON-NLS-1$
         String GENERATE_EXISTS = "prefGenerateExist"; //$NON-NLS-1$
+        String GENERATE_EXIST_DO_BLOCK = "prefGenerateExistDoBlock"; //$NON-NLS-1$
         String DROP_BEFORE_CREATE = "prefDropBeforeCreate"; //$NON-NLS-1$
         String COMMENTS_TO_END = "prefCommmentsToEnd"; //$NON-NLS-1$
     }